Change file structure and fix Makefile errors. Also had to squash bugs present with type mismatch

+++ lib/stats.c
@@ -3,7 +3,7 @@
 int32_t stats_sum(const int32_t *data, size_t n)
 {
     int32_t sum = 0;
-    for (int i = 0; i < n; i++)
+    for (size_t i = 0; i < n; i++)
     {
         sum = data[i] + sum;
     }
@@ -20,7 +20,7 @@ bool stats_min_max(const int32_t *data, size_t n,int32_t *out_min, int32_t *out_
     *out_min = data[0];
     *out_max = data[0];
 
-    for(int i = 0; i < n; i++)
+    for(size_t i = 0; i < n; i++)
     {
         if (data[i] < *out_min)
         {
